Extract the catalog service to own class
Up until now, CatalogBloc was acting as its own service. In a real-world app, BLoCs are composed of services that they call into, like Firestore and such.

With this change, all the logic in CatalogBloc has to do with converting streams and building `CatalogSlice`s. The logic of coming up with products has been moved to `CatalogService` with the intention to export that to other implementations.